backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-cache2>=0.2.2",
    "fastapi[standard]>=0.135.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "redis>=7.4.0",
    "ruff>=0.15.8",
]
